backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
langchain-aws
langchain-chroma
chromadb
boto3
pandas
python-dotenv